added header with enums and included it in various required locations
